Restructure CollectionDetailView: extract CollectionCoverView
Pulled the paged cover carousel (custom cover + recipe pages, page indicator,
and its own cover-image loading state) into a standalone CollectionCoverView.
CollectionDetailView drops from 1091 to 866 lines and no longer carries the
4 cover @State vars, the cover load function, or the page-clamp onChange.
Behavior-neutral; verified the cover renders/pages in the simulator.
